+++ com.unity.ml-agents/Runtime/Communicator/GrpcExtensions.cs
@@ -7,6 +7,8 @@ using UnityEngine;
 using System.Runtime.CompilerServices;
 using MLAgents.Sensors;
 using MLAgents.Demonstrations;
+using MLAgents.Policies;
+
 
 [assembly: InternalsVisibleTo("Unity.ML-Agents.Editor")]
 [assembly: InternalsVisibleTo("Unity.ML-Agents.Editor.Tests")]