+++ com.unity.ml-agents/Tests/Editor/MultinomialTest.cs
@@ -1,5 +1,5 @@
 using NUnit.Framework;
-using MLAgents.InferenceBrain.Utils;
+using MLAgents.Inference.Utils;
 
 namespace MLAgents.Tests
 {