+++ com.unity.ml-agents/Tests/Editor/ParameterLoaderTest.cs
@@ -2,9 +2,10 @@ using NUnit.Framework;
 using UnityEngine;
 using UnityEditor;
 using Barracuda;
-using MLAgents.InferenceBrain;
+using MLAgents.Inference;
 using MLAgents.Sensors;
 using System.Linq;
+using MLAgents.Policies;
 
 namespace MLAgents.Tests
 {